Refactor AI-related imports and structure
- Updated import paths for built-in AI modules from "@shared/built-in-ai" to "@shared/ai" across multiple files.
- Introduced new files for AI functionalities including use cases for LanguageDetector, LanguageModel, Proofreader, Rewriter, Summarizer, Translator, and Writer.
- Added download progress management and shared context handling for AI sessions.
- Implemented tests for the new AI spec and functionalities, ensuring proper session management and error handling.
- Removed deprecated useWriter from the old built-in AI structure.
