+++ src/shared/ai/built-in-ai/__fixtures__/translator-fake.ts
