+++ src/shared/ai/built-in-ai/spec.ts
