+++ src/shared/ai/built-in-ai/useBuiltInAI.test.tsx
