+++ src/shared/ai/sharedContext.ts
